Make tests work locally
