+++ .github/workflows/tests.yml
@@ -15,4 +15,4 @@ jobs:
         with:
           php_version: ${{ matrix.php }}
       - run: docker run -d -p 4222:4222 nats --js
-      - run: vendor/bin/phpunit tests
+      - run: vendor/bin/phpunit --testsuite Tests
